Methods in Spam service, reusable finally!
